# Performance optimization of LayerComposition Update