--- a/src/framework/components/particle-system/system.js
+++ b/src/framework/components/particle-system/system.js
@@ -236,7 +236,7 @@ Object.assign(ParticleSystemComponentSystem.prototype, {
                                 lightCube[i * 3 + 1] = this.app.scene.ambientLight.g;
                                 lightCube[i * 3 + 2] = this.app.scene.ambientLight.b;
                             }
-                            var dirs = layer._sortedLights[LIGHTTYPE_DIRECTIONAL];
+                            var dirs = layer._splitLights[LIGHTTYPE_DIRECTIONAL];
                             for (j = 0; j < dirs.length; j++) {
                                 for (c = 0; c < 6; c++) {
                                     var weight = Math.max(emitter.lightCubeDir[c].dot(dirs[j]._direction), 0) * dirs[j]._intensity;