Performance optimization of LayerComposition._update function, which executes when something in the scene changes (mesh / light / camera added / removed from layer, material blend mode changes ...)

Main reason for low performance - merging / searching arrays using array.IndexOf lookup. 
Solution - using Set / Map as needed, also reorganized code to optimize algorithms

Test case (on Mac)
1. Hierarchy engine example, with one mesh being added / removed from hierarchy each frame.
- test with 5000 mesh, performance improvement from 70ms to 2ms per frame
- test with 1300 meshes, performance improvement from 5ms to 0.1ms per frame

Expected impact:
- in applications with higher number of meshes in the scene, regularly changing the scene should be a lot faster.