@mvaligursky 2020-10-27
> Just to ensure: IE11 has limited support for features of Set and Map, does this PR only relies on supported features of IE11?

yes I only used few functions supported everywhere.