@mvaligursky 2020-10-28
> Did you consider wrapping the array+set combo into a reusable data structure, say `UniqueArray`?
Yes it crossed my mind but then decided against it. Changing (private) interface, extra layer to access things.